Update ADC block
Starting from R2013b or something, it is not possible to update block
'MaskValues' parameter from within a mask parameter callback. For some
reason it still works for 'MaskEnables' and 'MaskVisibilities'. So now i
work around this by passing the parameter name directly and using
set_param with this particular parameter, rather than the while
'MaskValues'.